feat: refactor PluginSchema
