+++ src/components/PluginForm/index.ts
@@ -1,0 +1,1 @@
+export {default} from './PluginForm'
