+++ src/components/PluginForm/service.ts
@@ -1,7 +1,7 @@
 import { request } from 'umi';
-import { transformSchemaFromAPI } from '@/transforms/plugin';
+import { transformSchemaFromAPI } from './transformer';
 
 export const fetchList = () => request('/api/plugins/list');
 
-export const fetchPluginSchema = (name: string): Promise<PluginSchema> =>
+export const fetchPluginSchema = (name: string): Promise<PluginForm.PluginSchema> =>
   request(`/schema/plugins/${name}`).then((data) => transformSchemaFromAPI(data, name));
