+++ src/typings.d.ts
@@ -38,46 +38,3 @@ declare let ANT_DESIGN_PRO_ONLY_DO_NOT_USE_IN_YOUR_PRODUCTION: 'site' | undefine
 declare const REACT_APP_ENV: 'test' | 'dev' | 'pre' | false;
 
 type PageMode = 'CREATE' | 'EDIT' | 'VIEW';
-
-interface PluginProperty {
-  type: 'number' | 'string' | 'integer' | 'array' | 'boolean' | 'object';
-  // the same as type
-  default?: any;
-  description?: string;
-  // NOTE: maybe 0.00001
-  minimum?: number;
-  maximum?: number;
-  minLength?: number;
-  maxLength?: number;
-  minItems?: number;
-  maxItems?: number;
-  // e.g "^/.*"
-  pattern?: string;
-  enum?: string[];
-  requried?: string[];
-  minProperties?: number;
-  additionalProperties?: boolean;
-  items?: {
-    type: string;
-    anyOf?: Array<{
-      type?: string;
-      description?: string;
-      enum?: string[];
-      pattern?: string;
-    }>;
-  };
-}
-
-interface PluginSchema {
-  type: 'object';
-  id?: string;
-  required?: string[];
-  additionalProperties?: boolean;
-  minProperties?: number;
-  oneOf?: Array<{
-    required: string[];
-  }>;
-  properties?: {
-    [propertyName: string]: PluginProperty;
-  };
-}
